fix #1 and #2
