+++ src/main/java/zone/rong/fluidizedtanks/block/TankBlock.java
@@ -1,7 +1,5 @@
 package zone.rong.fluidizedtanks.block;
 
-import net.minecraft.client.color.block.BlockColor;
-import net.minecraft.client.color.item.ItemColor;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.NonNullList;
@@ -36,7 +34,7 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 
-public class TankBlock extends Block implements EntityBlock, BlockColor, ItemColor {
+public class TankBlock extends Block implements EntityBlock {
 
     public TankBlock() {
         super(Properties.of(Material.BUILDABLE_GLASS).strength(1F).sound(SoundType.STONE).noOcclusion());
@@ -121,31 +119,6 @@ public class TankBlock extends Block implements EntityBlock, BlockColor, ItemCol
     @Override
     public boolean skipRendering(BlockState selfState, BlockState neighbourState, Direction direction) {
         return neighbourState.is(this) || super.skipRendering(selfState, neighbourState, direction);
-    }
-
-    @Override
-    @OnlyIn(Dist.CLIENT)
-    public int getColor(BlockState state, @Nullable BlockAndTintGetter level, @Nullable BlockPos pos, int tintIndex) {
-        if (tintIndex == 0) {
-            if (level != null && pos != null) {
-                if (level.getBlockEntity(pos) instanceof TankBlockEntity tank) {
-                    return tank.getTankDefinition().map(TankDefinition::colour).orElse(-1);
-                }
-            }
-        }
-        return -1;
-    }
-
-    @Override
-    @OnlyIn(Dist.CLIENT)
-    public int getColor(ItemStack stack, int tintIndex) {
-        if (tintIndex == 0) {
-            TankDefinition tankDefinition = TankDefinition.get(stack);
-            if (tankDefinition != null) {
-                return tankDefinition.colour();
-            }
-        }
-        return -1;
     }
 
     private static <E extends BlockEntity, A extends BlockEntity> BlockEntityTicker<A> createTickerHelper(BlockEntityType<A> p_152133_, BlockEntityType<E> p_152134_, BlockEntityTicker<? super E> p_152135_) {
